[refactor] selector 위치변경 사용컴포넌트에 반영
- 변경된 selector 위치 반영
